zombie animation. // bug : img == null

+++ src/ZombiePanel.java
@@ -29,8 +29,7 @@ public class ZombiePanel extends JPanel
     frame.addKeyListener(new KeyboardHandler());
     Player player = new Player();
     entityManager.add(player);
-
-    entityManager.add(new LineZombie(player, new Point2D.Float (240, 240)));
+    entityManager.add(new LineZombie(player, new Point2D.Float (40, 40)));
     entityManager.add(new House(32, 32));
     new Timer(1000 / Settings.frameRate, this::timerTick).start();
   }
